# docs/openapi/Kst.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Kst.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Returns liveness status of the backend process.",
        "operationId": "GetHealth",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Returns readiness status once initialization is complete.",
        "operationId": "GetReady",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/status": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Returns typed system status for the frontend walking skeleton.",
        "operationId": "GetSystemStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/refresh": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Triggers a refresh cycle across all registered data sources and returns the resulting status.",
        "operationId": "PostSystemRefresh",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Returns the saved workspace list and any nonfatal configuration warning.",
        "operationId": "ListWorkspaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponseDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Creates and persists a new workspace configuration.",
        "operationId": "CreateWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Removes all workspace configuration and returns to the empty startup state.",
        "operationId": "ResetWorkspaces",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/workspaces/{assignmentId}": {
      "put": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Updates and persists an existing workspace configuration.",
        "operationId": "UpdateWorkspace",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Permanently deletes a workspace assignment from configuration.",
        "operationId": "DeleteWorkspace",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{assignmentId}/archive": {
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Archives a workspace so it no longer appears as an active tab (isEnabled = false).",
        "operationId": "ArchiveWorkspace",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAssignmentDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{assignmentId}/restore": {
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Restores a previously archived workspace (isEnabled = true).",
        "operationId": "RestoreWorkspace",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAssignmentDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/order": {
      "put": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Persists a new tab order for the currently active (enabled) workspaces.",
        "operationId": "ReorderWorkspaces",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderWorkspacesRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "summary": "Returns the current user preferences and any nonfatal configuration warning.",
        "operationId": "GetPreferences",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Preferences"
        ],
        "summary": "Validates and persists updated user preferences.",
        "operationId": "UpdatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateWorkspaceRequestDto": {
        "required": [
          "displayName",
          "site",
          "customerNumber",
          "productLineFrom",
          "productLineTo",
          "isTemporary",
          "coverageEndsOn"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "customerNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "productLineFrom": {
            "type": [
              "null",
              "string"
            ]
          },
          "productLineTo": {
            "type": [
              "null",
              "string"
            ]
          },
          "isTemporary": {
            "type": "boolean"
          },
          "coverageEndsOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
      "DataSourceDto": {
        "required": [
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "HealthResponse": {
        "required": [
          "status",
          "application",
          "backendVersion",
          "processId",
          "instanceId",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "application": {
            "type": "string"
          },
          "backendVersion": {
            "type": "string"
          },
          "processId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "instanceId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PreferencesResponseDto": {
        "required": [
          "preferences",
          "configurationWarning"
        ],
        "type": "object",
        "properties": {
          "preferences": {
            "$ref": "#/components/schemas/UserPreferencesDto"
          },
          "configurationWarning": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReadyResponse": {
        "required": [
          "status",
          "initialized",
          "snapshotAvailable",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "initialized": {
            "type": "boolean"
          },
          "snapshotAvailable": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReorderWorkspacesRequestDto": {
        "required": [
          "assignmentIds"
        ],
        "type": "object",
        "properties": {
          "assignmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "SnapshotStatusDto": {
        "required": [
          "available",
          "snapshotId",
          "createdAt",
          "status"
        ],
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "snapshotId": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "SystemStatusResponse": {
        "required": [
          "applicationName",
          "applicationVersion",
          "backendFramework",
          "backendInstanceId",
          "startedAt",
          "currentTime",
          "snapshot",
          "dataSources",
          "lastRefreshAttemptAt",
          "lastSuccessfulRefreshAt"
        ],
        "type": "object",
        "properties": {
          "applicationName": {
            "type": "string"
          },
          "applicationVersion": {
            "type": "string"
          },
          "backendFramework": {
            "type": "string"
          },
          "backendInstanceId": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentTime": {
            "type": "string",
            "format": "date-time"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotStatusDto"
          },
          "dataSources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataSourceDto"
            }
          },
          "lastRefreshAttemptAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastSuccessfulRefreshAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "UpdatePreferencesRequestDto": {
        "required": [
          "theme",
          "accentColor",
          "rowDensity"
        ],
        "type": "object",
        "properties": {
          "theme": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "rowDensity": {
            "type": "string"
          }
        }
      },
      "UserPreferencesDto": {
        "required": [
          "theme",
          "accentColor",
          "rowDensity"
        ],
        "type": "object",
        "properties": {
          "theme": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "rowDensity": {
            "type": "string"
          }
        }
      },
      "WorkspaceAssignmentDto": {
        "required": [
          "assignmentId",
          "displayName",
          "site",
          "customerNumber",
          "productLineFrom",
          "productLineTo",
          "isTemporary",
          "coverageEndsOn",
          "isEnabled",
          "sortOrder"
        ],
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": "string"
          },
          "customerNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "productLineFrom": {
            "type": [
              "null",
              "string"
            ]
          },
          "productLineTo": {
            "type": [
              "null",
              "string"
            ]
          },
          "isTemporary": {
            "type": "boolean"
          },
          "coverageEndsOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "WorkspaceListResponseDto": {
        "required": [
          "workspaces",
          "configurationWarning"
        ],
        "type": "object",
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceAssignmentDto"
            }
          },
          "configurationWarning": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Diagnostics"
    },
    {
      "name": "System"
    },
    {
      "name": "Workspaces"
    },
    {
      "name": "Preferences"
    }
  ]
}